// x/ts/src/notation/types.gen.ts
// Copyright 2026 Synnax Labs, Inc.
//
// Use of this software is governed by the Business Source License included in the file
// licenses/BSL.txt.
//
// As of the Change Date specified in that file, in accordance with the Business Source
// License, use of this software will be governed by the Apache License, Version 2.0,
// included in the file licenses/APL.txt.

// Code generated by Oracle. DO NOT EDIT.

import { z } from "zod";

export const NOTATIONS = ["standard", "scientific", "engineering"] as const;
export const notationZ = z.enum(NOTATIONS);
export type Notation = z.infer<typeof notationZ>;